pyproject.toml
[project]
name = "livrar"
version = "0.1.0"
description = "Projeto Int"
authors = [
    {name = "Thigas", email = "thigas@gmail.com"},
]
dependencies = [
    "django>=4.2.6",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
